+++ noodles-vcf/src/async/io.rs
@@ -3,8 +3,4 @@
 mod reader;
 mod writer;
 
-#[deprecated(since = "0.79.0", note = "Use `vcf::r#async::io::Reader` instead.")]
-pub use self::reader::Reader;
-
-#[deprecated(since = "0.79.0", note = "Use `vcf::r#async::io::Writer` instead.")]
-pub use self::writer::Writer;
+pub use self::{reader::Reader, writer::Writer};