+++ noodles-vcf/src/lib.rs
@@ -27,11 +27,3 @@ pub mod record;
 pub mod variant;
 
 pub use self::{header::Header, record::Record};
-
-#[cfg(feature = "async")]
-#[deprecated(since = "0.79.0", note = "Use `vcf::r#async::io::Reader` instead.")]
-pub use self::r#async::io::Reader as AsyncReader;
-
-#[cfg(feature = "async")]
-#[deprecated(since = "0.79.0", note = "Use `vcf::r#async::io::Writer` instead.")]
-pub use self::r#async::io::Writer as AsyncWriter;